@@ -1,6 +1,6 @@
 package ru.vpcb.test.map.base;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public interface MvpPresenter<V extends MvpView> {
     void onAttach(@NonNull V view);
