@@ -1,6 +1,6 @@
 package ru.vpcb.test.map.login.signin;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.firebase.auth.FirebaseUser;
 
