@@ -7,9 +7,10 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.provider.Settings;
-import android.support.annotation.NonNull;
-import android.support.v4.content.LocalBroadcastManager;
-import android.support.v7.app.AlertDialog;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import com.google.android.gms.maps.CameraUpdate;
 import com.google.android.gms.maps.CameraUpdateFactory;
@@ -24,8 +25,8 @@ import java.util.List;
 
 import ru.vpcb.test.map.R;
 import ru.vpcb.test.map.data.provider.AddressLocationProvider;
-import ru.vpcb.test.map.executors.IListener;
 import ru.vpcb.test.map.data.provider.LocationProvider;
+import ru.vpcb.test.map.executors.IListener;
 import ru.vpcb.test.map.ext.PermissionExt;
 import ru.vpcb.test.map.home.HomeActivity;
 import ru.vpcb.test.map.model.Location;
