@@ -1,6 +1,7 @@
 package ru.vpcb.test.map.map;
 
-import android.support.v4.app.Fragment;
+
+import androidx.fragment.app.Fragment;
 
 public interface MapFragment {
 
