@@ -1,11 +1,13 @@
 package ru.vpcb.test.map.login.signin;
 
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.material.snackbar.Snackbar;
 
 import ru.vpcb.test.map.R;
 import ru.vpcb.test.map.data.repository.FirebaseUserRepository;
